Python/requirements.txt
openai>=1.0.0
pdf2image>=1.16.3
Pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0